Flytter samboer navn input ut fra ÅrsakEnslig og inn i OmDenTidligereSamboerenDin.tsx.
