+++ src/søknader/felles/steg/1-omdeg/sivilstatus/begrunnelse/ÅrsakEnslig.tsx
@@ -13,7 +13,6 @@ import { ISpørsmål, ISvar } from '../../../../../../models/felles/spørsmålog
 import { harFyltUtSamboerDetaljer } from '../../../../../../utils/person';
 import { useLokalIntlContext } from '../../../../../../context/LokalIntlContext';
 import { Alert, Heading } from '@navikt/ds-react';
-import { TextFieldMedBredde } from '../../../../../../components/TextFieldMedBredde';
 import { useOmDeg } from '../../OmDegContext';
 import { OmDenTidligereSamboerenDin } from './OmDenTidligereSamboerenDin';
 import { hentHTMLTekst, hentTekst } from '../../../../../../utils/teksthåndtering';
@@ -24,20 +23,6 @@ const ÅrsakEnslig: FC = () => {
   const { sivilstatus, settSivilstatus, settDokumentasjonsbehov } = useOmDeg();
 
   const { årsakEnslig, tidligereSamboerDetaljer } = sivilstatus;
-
-  const settNavn = (e: React.FormEvent<HTMLInputElement>) => {
-    settSivilstatus({
-      ...sivilstatus,
-      tidligereSamboerDetaljer: {
-        ...tidligereSamboerDetaljer,
-        kjennerIkkeIdent: tidligereSamboerDetaljer?.kjennerIkkeIdent ?? false,
-        navn: {
-          label: hentTekst('person.navn', intl),
-          verdi: e.currentTarget.value,
-        },
-      },
-    });
-  };
 
   const settÅrsakEnslig = (spørsmål: ISpørsmål, svar: ISvar) => {
     settSivilstatus({
@@ -79,16 +64,6 @@ const ÅrsakEnslig: FC = () => {
             <Heading size="small" level="3">
               {hentTekst('sivilstatus.tittel.samlivsbruddAndre', intl)}
             </Heading>
-          </FeltGruppe>
-          <FeltGruppe>
-            <TextFieldMedBredde
-              key={'navn'}
-              label={hentTekst('person.navn', intl)}
-              type="text"
-              bredde={'L'}
-              onChange={(e) => settNavn(e)}
-              value={tidligereSamboerDetaljer?.navn?.verdi}
-            />
           </FeltGruppe>
           <FeltGruppe>
             <OmDenTidligereSamboerenDin />
